code review feedback + reorganize the code a bit
